Fixed segfault by instantiating array in heap instead of stack.

+++ src/Grid.h
@@ -5,8 +5,8 @@
 #include <memory>
 
 struct WindowSize {
-    float width;
-    float height;
+    int width;
+    int height;
 };
 
 class Grid
